It (sorta) works!
At least I'm getting a boot message and a heartbeat from the node!

Had to force CONFIG_CANOPEN_STORAGE to n, and had to comment out the code in main() that attempted
to attach to storage and load from it.

I suspect that the FLASH EEPROM emulation for STM32L4xx is simply broken or non-existent.  Frankly
I'd be happy to forego CAN_STORAGE altogether, unless either we or the CANOpen spect *require*
non-volatile storage of dictionaries or node status.